Updated book borrow option, added pagination
Before, it would only show borrow button if a book is available.

now it shows how many copies of the book are available. and if the book is available there is a borrow button.

After clicking the borrow. the number of books available adjusts, and it is added to the users profile.